If the system is on a phone, a box is provided to allow the user to name the file. This is seamlessly done allowing a single click.
There were several issues with the session not
wanting to keep the file name. Once one was put in
and the download button was pressed, it would use
the previous filename, not the current one. This
is due to how streamlit handles the download
button. It doesn't cycle through the code again
before submitting and typing in text also doesn't
refresh the state. Custom html / js had to be
written to handle this.